2 - Extracted features from all images in template folder

+++ imageDetection.py
@@ -21,20 +21,24 @@ else:
     print("Error: No image detected")
 """
 directory = "templates"
-count = 1
-img = imread("templates/3_4 time signature.png")
-new_img = resize(img, (128*4, 64*4))
-if len(new_img.shape) == 3:
-    fd, hog_img = hog(new_img, orientations=9, pixels_per_cell=(8,8), cells_per_block=(2,2), visualize=True, channel_axis=-1)
-    plt.axis("off")
-    plt.imshow(hog_img, cmap="gray")
-    plt.savefig("template_features_output/feature {}.png".format(count))
-else:
-    fd, hog_img = hog(new_img, orientations=9, pixels_per_cell=(8,8), cells_per_block=(2,2), visualize=True)
-    plt.axis("off")
-    plt.imshow(hog_img, cmap="gray")
-    plt.savefig("template_features_output/feature {}.png".format(count))
-count += 1
+count = 0
+
+for filename in os.listdir(directory):
+    f = os.path.join(directory, filename)
+    if os.path.isfile(f):
+        img = imread(f)
+        new_img = resize(img, (128*4, 64*4))
+        if len(new_img.shape) == 3:
+            fd, hog_img = hog(new_img, orientations=9, pixels_per_cell=(8,8), cells_per_block=(2,2), visualize=True, channel_axis=-1)
+            plt.axis("off")
+            plt.imshow(hog_img, cmap="gray")
+            plt.savefig("template_features_output/feature {}.png".format(count))
+        else:
+            fd, hog_img = hog(new_img, orientations=9, pixels_per_cell=(8,8), cells_per_block=(2,2), visualize=True)
+            plt.axis("off")
+            plt.imshow(hog_img, cmap="gray")
+            plt.savefig("template_features_output/feature {}.png".format(count))
+        count += 1
 # testing HOG using skimage library
 
 """
